+++ src/gedi_subset/subset.py
@@ -7,24 +7,25 @@ import os.path
 from dataclasses import dataclass
 from enum import Enum
 from pathlib import Path
-from typing import Any, Iterable, Sequence, Tuple
+from typing import Any, Iterable, Mapping, Optional, Sequence, Tuple
 
 import geopandas as gpd
 import h5py
 import typer
 from maap.maap import MAAP
-from maap.Result import Granule
+from maap.Result import Collection, Granule
 from returns.curry import partial
 from returns.functions import raise_exception, tap
 from returns.io import IOFailure, IOResult, IOResultE, IOSuccess, impure_safe
 from returns.iterables import Fold
 from returns.maybe import Maybe, Nothing, Some
 from returns.pipeline import flow, is_successful, pipe
-from returns.pointfree import bind, bind_ioresult, lash, map_
+from returns.pointfree import bind, bind_ioresult, bind_result, lash, map_
+from returns.result import Failure, Success
 from returns.unsafe import unsafe_perform_io
 
+import gedi_subset.fp as fp
 from gedi_subset import osx
-from gedi_subset.fp import always, filter, map
 from gedi_subset.gedi_utils import (
     chext,
     gdf_read_parquet,
@@ -40,6 +41,18 @@ class CMRHost(str, Enum):
     maap = "cmr.maap-project.org"
     nasa = "cmr.earthdata.nasa.gov"
 
+    def __str__(self) -> str:
+        return self.value
+
+
+logical_dois = {
+    "L1B": "10.5067/GEDI/GEDI01_B.002",
+    "L2A": "10.5067/GEDI/GEDI02_A.002",
+    "L2B": "10.5067/GEDI/GEDI02_B.002",
+    "L4A": "10.3334/ORNLDAAC/2056",
+}
+
+DEFAULT_LIMIT = 10_000
 
 LOGGING_FORMAT = "%(asctime)s [%(processName)s:%(name)s] [%(levelname)s] %(message)s"
 
@@ -60,9 +73,50 @@ class SubsetGranuleProps:
     granule: Granule
     maap: MAAP
     aoi_gdf: gpd.GeoDataFrame
+    lat: str
+    lon: str
     columns: Sequence[str]
-    query: str
+    query: Optional[str]
     output_dir: Path
+
+
+def is_gedi_collection(c: Collection) -> bool:
+    """Return True if the specified collection is a GEDI collection containing granule
+    data files in HDF5 format; False otherwise."""
+
+    c = c.get("Collection", {})
+    attrs = c.get("AdditionalAttributes", {}).get("AdditionalAttribute", [])
+    data_format_attrs = (attr for attr in attrs if attr.get("Name") == "Data Format")
+    data_format = next(data_format_attrs, {"Value": c.get("DataFormat")}).get("Value")
+
+    return c.get("ShortName", "").startswith("GEDI") and data_format == "HDF5"
+
+
+def find_gedi_collection(
+    maap: MAAP,
+    cmr_host: str,
+    params: Mapping[str, str],
+) -> IOResultE[Collection]:
+    """Find a GEDI collection matching the given parameters.
+
+    Return `IOSuccess[Collection]` containing the collection upon successful
+    search; otherwise return `IOFailure[Exception]` containing the reason for
+    failure, which is a `ValueError` when there is no matching collection or
+    the collection is _not_ a GEDI collection.
+    """
+    return flow(
+        find_collection(maap, cmr_host, params),
+        bind_result(
+            lambda c: Success(c)
+            if is_gedi_collection(c)
+            else Failure(
+                ValueError(
+                    f"Collection {c['Collection']['ShortName']} is not a GEDI"
+                    " collection, or does not contain HDF5 data files."
+                )
+            )
+        ),
+    )
 
 
 @impure_safe
@@ -74,8 +128,10 @@ def subset_granule(props: SubsetGranuleProps) -> Maybe[str]:
     file where it overlaps with the specified AOI (`props.aoi_gdf`), remove the
     downloaded granule file, and return the path to the output file.
 
-    Return `Nothing` if the subset is empty (in which case no GeoParquet file
-    was written), otherwise `Some[str]` indicating the output path of the
+    Return `Nothing` if the subset is empty or if an error occurred attempting
+    to read the granule file (in which case, the offending file is retained in
+    order to facilitate analysis).  In either case, no GeoParquet file is
+    written; otherwise return `Some[str]` indicating the output path of the
     GeoParquet file.
     """
 
@@ -84,8 +140,19 @@ def subset_granule(props: SubsetGranuleProps) -> Maybe[str]:
 
     logger.debug(f"Subsetting {inpath}")
 
-    with h5py.File(inpath) as hdf5:
-        gdf = subset_hdf5(hdf5, props.aoi_gdf, props.columns, props.query)
+    try:
+        hdf5 = h5py.File(inpath)
+    except Exception as e:
+        granule_ur = props.granule["Granule"]["GranuleUR"]
+        logger.warning(f"Skipping granule {granule_ur} [failed to read {inpath}: {e}]")
+        return Nothing
+
+    try:
+        gdf = subset_hdf5(
+            hdf5, props.aoi_gdf, props.lat, props.lon, props.columns, props.query
+        )
+    finally:
+        hdf5.close()
 
     osx.remove(inpath)
 
@@ -112,8 +179,10 @@ def set_logging_level(logging_level: int) -> None:
 def subset_granules(
     maap: MAAP,
     aoi_gdf: gpd.GeoDataFrame,
+    lat: str,
+    lon: str,
     columns: Sequence[str],
-    query: str,
+    query: Optional[str],
     output_dir: Path,
     dest: Path,
     init_args: Tuple[Any, ...],
@@ -133,15 +202,15 @@ def subset_granules(
         return flow(
             gdf_read_parquet(src),
             bind_ioresult(partial(gdf_to_file, dest, to_file_props)),
-            tap(pipe(always(src), osx.remove)),
-            map_(always(src)),
+            tap(pipe(fp.always(src), osx.remove)),
+            map_(fp.always(src)),
         )
 
     # https://docs.python.org/3/library/multiprocessing.html#multiprocessing.pool.Pool.imap
     chunksize = 10
     processes = os.cpu_count()
     payloads = (
-        SubsetGranuleProps(granule, maap, aoi_gdf, columns, query, output_dir)
+        SubsetGranuleProps(granule, maap, aoi_gdf, lat, lon, columns, query, output_dir)
         for granule in granules
     )
 
@@ -150,9 +219,9 @@ def subset_granules(
     with multiprocessing.Pool(processes, init_process, init_args) as pool:
         return flow(
             pool.imap_unordered(subset_granule, payloads, chunksize),
-            map(lash(raise_exception)),  # Fail fast (if subsetting errored out)
-            filter(subset_saved),  # Skip granules that produced empty subsets
-            map(bind(bind(append_subset))),  # Append non-empty subset
+            fp.map(lash(raise_exception)),  # Fail fast (if subsetting errored out)
+            fp.filter(subset_saved),  # Skip granules that produced empty subsets
+            fp.map(bind(bind(append_subset))),  # Append non-empty subset
             partial(Fold.collect, acc=IOSuccess(())),
         )
 
@@ -169,35 +238,35 @@ def main(
         resolve_path=True,
     ),
     doi=typer.Option(
-        "10.3334/ORNLDAAC/2056",  # GEDI L4A DOI, v2.1
-        help="Digital Object Identifier of collection to subset (https://www.doi.org/)",
+        ...,
+        callback=lambda value: logical_dois.get(value.upper(), value),
+        help=(
+            "Digital Object Identifier (DOI) of collection to subset"
+            " (https://www.doi.org/), or one of these logical, case-insensitive"
+            f" names: {', '.join(logical_dois)}"
+        ),
     ),
     cmr_host: CMRHost = typer.Option(
         CMRHost.maap,
         help="CMR hostname",
     ),
+    lat: str = typer.Option(
+        ..., help=("Latitude dataset used in the geometry of the dataframe")
+    ),
+    lon: str = typer.Option(
+        ..., help=("Longitude dataset used in the geometry of the dataframe")
+    ),
     columns: str = typer.Option(
-        ",".join(
-            [
-                "agbd",
-                "agbd_se",
-                "l2_quality_flag",
-                "l4_quality_flag",
-                "sensitivity",
-                "sensitivity_a2",
-            ]
-        ),
+        ...,
         help="Comma-separated list of columns to select",
     ),
     query: str = typer.Option(
-        "l2_quality_flag == 1"
-        " and l4_quality_flag == 1"
-        " and sensitivity > 0.95"
-        " and sensitivity_a2 > 0.95",
+        None,
         help="Boolean query expression to select rows",
     ),
     limit: int = typer.Option(
-        10_000,
+        DEFAULT_LIMIT,
+        callback=lambda value: DEFAULT_LIMIT if value < 1 else value,
         help="Maximum number of granules to subset",
     ),
     output_dir: Path = typer.Option(
@@ -231,22 +300,27 @@ def main(
     IOResult.do(
         subsets
         for aoi_gdf in impure_safe(gpd.read_file)(aoi)
-        for collection in find_collection(maap, cmr_host, {"doi": doi})
+        # Use wildcards around DOI value because some collections have incorrect
+        # DOI values. For example, the L2B collection has the full DOI URL as
+        # the DOI value (i.e., https://doi.org/<DOI> rather than just <DOI>).
+        for collection in find_gedi_collection(maap, cmr_host, {"doi": f"*{doi}*"})
         for granules in impure_safe(maap.searchGranule)(
             cmr_host=cmr_host,
             collection_concept_id=collection["concept-id"],
-            bounding_box=",".join(map(str)(aoi_gdf.total_bounds)),
+            bounding_box=",".join(fp.map(str)(aoi_gdf.total_bounds)),
             limit=limit,
         )
         for subsets in subset_granules(
             maap,
             aoi_gdf,
+            lat,
+            lon,
             [c.strip() for c in columns.split(",")],
             query,
             output_dir,
             dest,
             (logging_level,),
-            filter(partial(granule_intersects, aoi_gdf.geometry[0]))(granules),
+            fp.filter(partial(granule_intersects, aoi_gdf.geometry[0]))(granules),
         )
     ).bind_ioresult(
         lambda subsets: IOSuccess(subsets)